+++ four/src/iam/property/access_key.rs
@@ -1,0 +1,13 @@
+use serde::Serialize;
+
+#[derive(Debug, Clone, Serialize)]
+pub enum AccessKeyStatus {
+    Inactive,
+    Active,
+}
+
+#[derive(Debug)]
+pub struct AccessKeyId;
+
+#[derive(Debug, Clone, Serialize)]
+pub struct SecretAccessKey(String);
