+++ four/src/iam/property/group.rs
@@ -1,0 +1,6 @@
+use serde::Serialize;
+
+use crate::{service, Arn};
+
+#[derive(Debug, Clone, Serialize)]
+pub struct GroupArn(Arn<service::IAM>);
