+++ four/src/iam/property/group_policy.rs
@@ -1,0 +1,10 @@
+use nutype::nutype;
+
+#[nutype(
+    validate(not_empty, len_char_max = 128, regex = r#"[\w+=,.@-]+"#),
+    derive(Debug, Clone, Serialize)
+)]
+pub struct GroupName(String);
+
+#[derive(Debug)]
+pub struct GroupPolicyId;
