+++ four/src/iam/property/id.rs
@@ -1,0 +1,2 @@
+#[derive(Debug)]
+pub struct InstanceProfileId;
