+++ four/src/iam/property/instance_profile.rs
@@ -1,0 +1,7 @@
+use nutype::nutype;
+
+#[nutype(
+    validate(not_empty, len_char_max = 128, regex = r#"[\w+=,.@-]+"#),
+    derive(Debug, Clone, Serialize)
+)]
+pub struct InstanceProfileName(String);
