+++ four/src/iam/resource/group_policy.rs
@@ -4,10 +4,12 @@ use crate::{
         function::{RefInner, Referenced},
         LogicalId,
     },
-    iam::{property::policy_document::PolicyDocument, resource::policy::PolicyName},
+    iam::{
+        property::{policy_document::PolicyDocument, GroupName, GroupPolicyId},
+        resource::policy::PolicyName,
+    },
 };
 use four_derive::ManagedResource;
-use nutype::nutype;
 
 #[derive(ManagedResource, Clone)]
 #[resource_type = "AWS::IAM::GroupPolicy"]
@@ -17,15 +19,6 @@ pub struct GroupPolicy {
     policy_document: Option<PolicyDocument>,
     policy_name: WillBe<PolicyName>,
 }
-
-#[nutype(
-    validate(not_empty, len_char_max = 128, regex = r#"[\w+=,.@-]+"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct GroupName(String);
-
-#[derive(Debug)]
-pub struct GroupPolicyId;
 
 impl Referenced for GroupPolicy {
     type To = GroupPolicyId;
