+++ four/src/iam/resource/instance_profile.rs
@@ -1,15 +1,16 @@
 use crate::{
     core::{
-        convert::{WillBe, WillMappable},
+        convert::WillBe,
         function::{HaveAtt, RefInner, Referenced},
-        service::IAM,
-        Arn, LogicalId,
+        LogicalId,
     },
-    iam::{resource::role::RoleName, util::Path},
+    iam::{
+        property::{InstanceProfileArn, InstanceProfileId, InstanceProfileName},
+        resource::role::RoleName,
+        util::Path,
+    },
 };
 use four_derive::ManagedResource;
-use nutype::nutype;
-use serde::Serialize;
 
 #[derive(ManagedResource, Clone)]
 #[resource_type = "AWS::IAM::InstanceProfile"]
@@ -20,17 +21,6 @@ pub struct InstanceProfile {
     roles: Vec<WillBe<RoleName>>,
 }
 
-#[nutype(
-    validate(not_empty, len_char_max = 128, regex = r#"[\w+=,.@-]+"#),
-    derive(Debug, Clone, Serialize)
-)]
-pub struct InstanceProfileName(String);
-
-impl WillMappable<String> for InstanceProfileName {}
-
-#[derive(Debug)]
-pub struct InstanceProfileId;
-
 impl Referenced for InstanceProfile {
     type To = InstanceProfileId;
 
@@ -38,9 +28,6 @@ impl Referenced for InstanceProfile {
         RefInner::Id(self.logical_id.clone())
     }
 }
-
-#[derive(Debug, Clone, Serialize)]
-pub struct InstanceProfileArn(Arn<IAM>);
 
 impl HaveAtt<InstanceProfileArn> for InstanceProfile {
     const KEY: &'static str = "Arn";
